chore: renaming of functions and files
