+++ tests/Dfe.PlanTech.Web.E2ETests/cypress/e2e/dynamic-page-validator/validators/index.js
@@ -3,11 +3,11 @@ export { validateCheckAnswersPage } from "./check-answers.js";
 export { validateCompletionTags } from "./completion-tags.js";
 export { validateNavigationLinks } from "./nav-links.js";
 export { validateAndTestNonAuthorisedPages } from "./non-auth-pages.js";
-export { navigateAndValidateQuestionPages } from "./questions.js";
+export { validateQuestionPages } from "./validate-questions.js";
 export { validateRecommendationChunks } from "./recommendation-chunks.js";
 export { validateRecommendationIntro } from "./recommendation-intro.js";
-export { validateRecommendationForMaturity } from "./recommendations.js";
+export { validateAndTestRecommendations } from "./validate-recommendations.js";
 export { validateSelfAssessmentPage } from "./self-assessment-page.js";
-export { validateAndTestSections } from "./section.js";
-export { quickNavigateToRecommendations } from "./quick-navigate-recommendations.js";
-export { extraQuestionTests } from "./questions-extra.js";
+export { validateAndTestSections } from "./validate-section.js";
+export { minimalSectionValidationForRecommendations } from "./minimal-validation-recommendations.js";
+export { validateAnswersHintAndUrl } from "./validate-questions-additional.js";
